Add two additional index implementations.
This commit extends the previous "constexpr vector-like" index implementation
with a fixed-size index implementation and a sequence-style implementation, and
updates the infrastructure so that the client and library code works with any of
the three depending on an `inline namespace` selection.

The fixed-size index implementation just sets a fixed maximum size for the index
array of 8, which means that trees must not contain anything over 7th order
tensor expressions. This is more than adequate for the current
infrastructure. It provides the same semantic power as the vector-like version
but compiles faster (as it produces many fewer types and thus symbols and binary
size).

The sequence-style index implementation encodes indices in compile-time
sequence types (like integer_sequence). It encodes the index sets in the type of
the tree rather than as values, and can make certain constexpr code operations
easier, but is the costliest to compile and creates the largest binaries.