Remove the postorder traversal from the tree.
It turns out that doing the postorder traversal is so easy that it really
doesn't help all that much to abstract it away, and the point of use ends up
looking more complex than it needs to.

This is somewhat governed by the fact that we have the tag-based nodes, rather
than variants.