Extend the system with code to simplify trees.
Tree simplification means performing symbolic algebra and calculus in order to
transform the tree in meaningful ways.

Of primary importance to us is simplifying derivatives, because we do not want
to have to synthesize derivatives of intermediate state while integrating. The
simplifier knows how to distribute over {+,-}, and perform the product and
quotient rules. It also knows that the derivative of a constant expression is
0.

Simplification is performed on a dynamically-allocated, pointer-based tree
structure. This structure cannot appear in top-level constexpr context until we
can leak dynamically allocated data out of constexpr evaluation.