Encapsulate ScalarTreeBuilder infrastructure into the ScalarTree.
This restructures the ScalarTree to be an outer interface for the linked tree
structure. The Tree will own its nodes and memory manage them on its own so that
we don't need to use the box structure, which we can now entirely eliminate.

We retain the builder structure because it is responsible for enumerating the
outer dimensions of the tensor tree, and storing the resulting scalar trees in
the external vector.