Add an experimental EveTree that encodes the tree geometry in the tree type.
This implementation doesn't require the compiler to #pragma unroll anything in
order to evaluate the tree to see through the runtime switch statement. Instead,
it encodes the tree topology (Tag types) in the executable tree type, and uses a
compile-time fold operation and `if constexpr` in order to expand the stack
machine.

It produces straightline code for all of the Sedov example systems, and should
also allow us to embed a set of SIMD-aware functions into the tree (exp, pow,
trig, etc).