Checkpoint work on transitioning to CNTTP.
This allows the serialized tensor tree to be used directly as CNTTP for the
executable tensor tree, however disables immediates for the time being because
they're `double` and that's not directly supported until gcc-11 (and not at all
in any clang), and there's no bit_cast in gcc-10.

The next step will be to build the immediates independently and store them as a
variable.